Pipe the event through DropDownMenu#onToggle to onClose and onOpen
DropDownMenu#onToggle does not pipe the click event through to
onClose and onOpen.  Without the event consumers of DropDownMenu
are unable to respond appropriately.

This commit also addresses several other cases of the event not
being piped through and corrects a bug in handleKeyDown resulting
in the handleClose callback being executed before setState finishes